get_xml_buffer: fix signed/unsigned error on Wii
With powerpc-eabi-gcc 4.5.1 (devkitPPC Release 21) the following error is
produced when using a ternary operator where the arguments are of different
types:

    get_xml_buffer.c:24:22: error: signed and unsigned type
        in conditional expression

This fix makes explicit what I think is an implicit cast that would be done
without using the ternary operator.

+++ libiview/get_xml_buffer.c
@@ -21,7 +21,9 @@ ssize_t iv_get_xml_buffer(const char *uri, char **buf_ptr) {
     res = http_request(uri, 1 << 31);
     http_get_result(&status, (u8 **)buf_ptr, &len);
     IV_DEBUG("Request %d, HTTP status %d, length %d.\n", res, status, len);
-    return res ? len :-IV_EREQUEST;
+    /* Cast is to avoid an error with powerpc-eabi-gcc.  Should be safe for any
+     * sane xml buffer length.*/
+    return res ? (ssize_t)len :-IV_EREQUEST;
 }
 #else
     int return_val;
